init.lua: deletes autoformat
NEVER EVER format on save!! Kill it with fire!

I could just have edit the `format_on_save` option, but I am not sure
to understand how to correctly do it, so delete and forget.
IMHO it’s also a better practice to explicitly use external tools so we
can apprehend them better.

Signed-off-by: Thibaud CANALE <thican@thican.net>